Move test only dummy data to test file
