+++ src/pages/library/datasetBuilder/DomainCriteriaSelector.test.ts
@@ -2,12 +2,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import _ from 'lodash/fp';
 import { h } from 'react-hyperscript-helpers';
-import {
-  DatasetBuilder,
-  DatasetBuilderContract,
-  dummyDatasetDetails,
-  getConceptForId,
-} from 'src/libs/ajax/DatasetBuilder';
+import { DatasetBuilder, DatasetBuilderContract, getConceptForId } from 'src/libs/ajax/DatasetBuilder';
 import {
   cohortEditorState,
   domainCriteriaSelectorState,
@@ -15,6 +10,7 @@ import {
   newCriteriaGroup,
 } from 'src/pages/library/datasetBuilder/dataset-builder-types';
 import { DomainCriteriaSelector, toCriteria } from 'src/pages/library/datasetBuilder/DomainCriteriaSelector';
+import { dummyDatasetDetails } from 'src/pages/library/datasetBuilder/TestConstants';
 import { asMockedFn } from 'src/testing/test-utils';
 
 type DatasetBuilderExports = typeof import('src/libs/ajax/DatasetBuilder');
